// src/types/mjo-card.d.ts
// Card contrast types
export type MjoCardContrast = "low" | "high" | "normal";
export type MjoCardRadius = "none" | "small" | "medium" | "large";
export type MjoCardSpace = "xxsmall" | "xsmall" | "small" | "medium" | "large" | "xlarge" | "xxlarge";
export type MjoCardVariants = "default" | "modern" | "skew";